:truck: : move inline function whenever in specific utils file
