+++ src/test/java/io/codeka/gaia/test/KotlinUtils.kt
@@ -1,0 +1,8 @@
+package io.codeka.gaia.test
+
+import org.mockito.Mockito
+import org.mockito.stubbing.OngoingStubbing
+
+inline fun <T> whenever(methodCall: T): OngoingStubbing<T> {
+    return Mockito.`when`(methodCall)!!
+}
